feat: update cli version

+++ cmd_build.go
@@ -45,7 +45,7 @@ func cmdBuild() *cli.Command {
 			FLAG_DRYRUN,
 		},
 		Action: func(ctx context.Context, cmd *cli.Command) error {
-			logger := getLogger("build", getLogLevel(cmd.Int("verbose")))
+			logger := getLogger("build", getLogLevel(cmd.Int64("verbose")))
 			slog.SetDefault(logger)
 
 			config, _ := loadConfig(cmd)
